requirements.txt
Appium-Python-Client
google-generativeai
selenium
python-dotenv